factor painter into separate file
