+++ lib/widgets/paper_widget.dart
@@ -1,32 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:sketchnotes_flutter/models/pen.dart';
+import 'package:sketchnotes_flutter/widgets/paper_painter.dart';
 import 'package:sketchnotes_flutter/widgets/user_preferences_provider.dart';
-
-class PaperPainter extends CustomPainter {
-  PaperPainter(this.points, this.color);
-
-  final List<Offset> points;
-  final Color color;
-
-  void paint(Canvas canvas, Size size) {
-    Paint paint = new Paint()
-      ..color = color
-      ..strokeCap = StrokeCap.round
-      ..strokeWidth = 5.0;
-    for (int i = 0; i < points.length - 1; i++) {
-      if (points[i] != null && points[i + 1] != null)
-        canvas.drawLine(points[i], points[i + 1], paint);
-    }
-  }
-
-  bool shouldRepaint(PaperPainter other) => other.points != points;
-}
 
 class PaperWidget extends StatefulWidget {
   final _state = PaperWidgetState();
 
   PaperWidgetState createState() => _state;
 
+  // TODO: this needs to come from Bloc emitting new-page event stream
   void clear() {
     _state.clear();
   }
